New CommonUI version (Still a lot WIP)
Finally, here is the promised new version using CommonUI. Now it's a full game menu system with an insane amount of settings. It's a back-end dream for all UI developper which will still be improved a lot. Everything is reusable and easy to understand. A documentation will come as well as a lot of comment in the project Blueprint.
Enjoy! But keep in mind that it's not finished at all and a lot will come!