Fixed Critcal Bug!!!!!
Fixed the problem when launching the project for the first time, where it was attempting to load a savegame that did not yet exist, causing each value to be set to 0.
The worst thing was that I did something that checked whether it existed or not in the load function but if it didn't exist it didn't return anything so it still executed what preceded the return node so if the savegame existed, and it was causing a lot of errors because of invalid savegame reference.
And the fix was so easy ;-;